+++ grid2op/tests/test_resest_options.py
@@ -65,7 +65,7 @@ class InitTSOptions(unittest.TestCase):
         assert obs.hour_of_day == 0, f"{ obs.hour_of_day} vs 1"
         assert obs.minute_of_hour == 0, f"{ obs.minute_of_hour} vs 0"
     
-    def check_soft_overflow(self):
+    def test_soft_overflow(self):
         """check that the lines are not on soft overflow (obs.timestep_overflow == 0 just after reset)"""
         line_id = 3
         obs = self.env.reset(options={"time serie id": 0})
@@ -97,7 +97,7 @@ class InitTSOptions(unittest.TestCase):
         assert obs.rho[line_id] > 1.
         assert obs.line_status[line_id]
     
-    def check_hard_overflow(self):
+    def test_hard_overflow(self):
         """check lines are disconnected if on hard overflow at the beginning"""
         line_id = 3
         obs = self.env.reset(options={"time serie id": 0})
@@ -142,10 +142,150 @@ class InitTSOptions(unittest.TestCase):
         with self.assertRaises(Grid2OpException):
             # float which is not an int
             obs = self.env.reset(options={"init ts": 1.5})
+        with self.assertRaises(Grid2OpException):
+            # value too small
+            obs = self.env.reset(options={"init ts": 0})
             
         # should work with a float convertible to an int
         obs = self.env.reset(options={"time serie id": 0, "init ts": 6.})
     
+
+class MaxStepOptions(unittest.TestCase):
+    """test the "max step" options in env.reset() """
+    def setUp(self) -> None:
+        self.env_name = "l2rpn_case14_sandbox"
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore")
+            self.env = grid2op.make(self.env_name, test=True,
+                                    _add_to_name=type(self).__name__)
+            
+    def test_raise_if_args_not_correct(self):
+        with self.assertRaises(Grid2OpException):
+            # string and not int
+            obs = self.env.reset(options={"max step": "treliug"})
+        with self.assertRaises(Grid2OpException):
+            # float which is not an int
+            obs = self.env.reset(options={"max step": 1.5})
+            
+        with self.assertRaises(Grid2OpException):
+            # value too small
+            obs = self.env.reset(options={"max step": 0})
+            
+        # should work with a float convertible to an int
+        obs = self.env.reset(options={"time serie id": 0, "max step": 6.})
+
+    def test_function_ok(self):
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+        
+        # enough data to be limited
+        obs = self.env.reset(options={"max step": 5})
+        assert obs.max_step == 5, f"{obs.max_step} vs 5"
     
+        # limit has no effect: not enough data anyway
+        obs = self.env.reset(options={"max step": 800})
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+    
+    def test_no_impact_next_reset(self):
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+        
+        # enough data to be limited
+        obs = self.env.reset(options={"max step": 5})
+        assert obs.max_step == 5, f"{obs.max_step} vs 5"
+        
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+    
+    def test_remember_previous_max_iter(self):
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+        
+        self.env.set_max_iter(200)
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 200, f"{obs.max_step} vs 200"
+        
+        # use the option to limit
+        obs = self.env.reset(options={"max step": 5})
+        assert obs.max_step == 5, f"{obs.max_step} vs 5"
+        
+        # check it remembers the previous limit
+        obs = self.env.reset() # normal reset (but 200 were set)
+        assert obs.max_step == 200, f"{obs.max_step} vs 200"
+        
+        # set back the limit to "maximum in the time serie"
+        self.env.set_max_iter(-1)
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+        
+        # limit for this reset only
+        obs = self.env.reset(options={"max step": 5})
+        assert obs.max_step == 5, f"{obs.max_step} vs 5"
+        
+        # check again the right limit was applied
+        obs = self.env.reset() # normal reset (but 575 were set back)
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+    
+    def test_max_step_and_init_ts(self):
+        """test that episode duration is properly computed and updated in 
+        the observation when both max step and init ts are set at the same time"""
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+        assert obs.year == 2019
+        assert obs.month == 1
+        assert obs.day == 6
+        assert obs.hour_of_day == 0
+        assert obs.minute_of_hour == 0  
+        
+        obs = self.env.reset(options={"init ts": 12 * 24, "max step": 24})  # start after exactly 1 day for 2 hours
+        assert obs.year == 2019
+        assert obs.month == 1
+        assert obs.day == 7, f"{ obs.day} vs 7"
+        assert obs.hour_of_day == 0, f"{ obs.hour_of_day} vs 1"
+        assert obs.minute_of_hour == 0, f"{ obs.minute_of_hour} vs 0"
+        assert obs.max_step == 24, f"{obs.max_step} vs 24"
+        
+        obs = self.env.reset(options={"init ts": 12 * 24})  # start after exactly 1 day  without any max
+        assert obs.year == 2019
+        assert obs.month == 1
+        assert obs.day == 7, f"{ obs.day} vs 7"
+        assert obs.hour_of_day == 0, f"{ obs.hour_of_day} vs 1"
+        assert obs.minute_of_hour == 0, f"{ obs.minute_of_hour} vs 0"
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+        
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+        assert obs.year == 2019
+        assert obs.month == 1
+        assert obs.day == 6
+        assert obs.hour_of_day == 0
+        assert obs.minute_of_hour == 0  
+        
+        obs = self.env.reset(options={"max step": 288})  # don't skip anything, but last only 1 day
+        assert obs.year == 2019
+        assert obs.month == 1
+        assert obs.day == 6, f"{ obs.day} vs 6"
+        assert obs.hour_of_day == 0, f"{ obs.hour_of_day} vs 1"
+        assert obs.minute_of_hour == 0, f"{ obs.minute_of_hour} vs 0"
+        assert obs.max_step == 288, f"{obs.max_step} vs 288"
+        
+        obs = self.env.reset(options={"init ts": 12 * 24, "max step": 700})  # start after exactly 1 day for too much steps
+        assert obs.year == 2019
+        assert obs.month == 1
+        assert obs.day == 7, f"{ obs.day} vs 7"
+        assert obs.hour_of_day == 0, f"{ obs.hour_of_day} vs 1"
+        assert obs.minute_of_hour == 0, f"{ obs.minute_of_hour} vs 0"
+        # 288 here because the limit is the time series !
+        assert obs.max_step == 287, f"{obs.max_step} vs 287"
+        
+        obs = self.env.reset() # normal reset
+        assert obs.max_step == 575, f"{obs.max_step} vs 575"
+        assert obs.year == 2019
+        assert obs.month == 1
+        assert obs.day == 6
+        assert obs.hour_of_day == 0
+        assert obs.minute_of_hour == 0  
+        
+        
 if __name__ == "__main__":
     unittest.main()