+++ grid2op/typing_variables.py
@@ -45,6 +45,8 @@ DICT_ACT_TYPING = Dict[Literal["set_line_status",
 #: type hints for the "options" flag of reset function
 RESET_OPTIONS_TYPING = Union[Dict[Literal["time serie id"], int],
                              Dict[Literal["init state"], DICT_ACT_TYPING],
+                             Dict[Literal["init ts"], int],
+                             Dict[Literal["max step"], int],
                              None]
 
 #: type hints for a "GridObject" when converted to a dictionary